server-docker/requirements.txt
fastmcp
python-dotenv

